+++ backend/src/backend/__main__.py
@@ -1,0 +1,5 @@
+"""convenience runner: `python -m backend`."""
+
+import uvicorn
+
+uvicorn.run("backend.main:app", reload=True)
